[Refactor] #233 - 유즈케이스 삭제 및 레포지토리 메서드 네이밍 변경
